See issue cited for master branch
The original code doesn't always work for 12:00-12:59am/pm. For example, if you use this code as is to find the number of minutes between "3:00pm-12:05am" you will get 1265 minutes for your answer because the program will think that it's counting from 3:00pm until 12:05pm the next day. The answer should be 545 minutes. Exceptions need to be made for the noon and midnight hours because these two hours uniquely spill over into the next am/pm cycle while still counting upward from the previous cycle. This is no longer an issue when 1:00(am or pm) is reached.

See my edits that make exceptions for the 12:00 hours to fix this problem.